requirements.txt
anyio==3.7.1
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
nmslib==2.1.1
numpy==1.25.2
opencv-python==4.8.0.76
psutil==5.9.5
pybind11==2.6.1
python-telegram-bot==20.4
sniffio==1.3.0
